--HG--
branch : soc.2013.gobjectification.plugins
rename : libpurple/protocols/mxit/protocol.c => libpurple/protocols/mxit/client.c
rename : libpurple/protocols/yahoo/libymsg.c => libpurple/protocols/yahoo/ymsg.c

+++ libpurple/protocols/msn/tlv.c
@@ -326,7 +326,7 @@ msn_tlvlist_remove(GSList **list, const guint8 type)
 }
 
 char *
-msn_tlvlist_write(GSList *list, size_t *out_len)
+msn_tlvlist_write(GSList *list, guint8 *out_len)
 {
 	char *buf;
 	char *tmp;
